requirements.txt
torch
torchvision
numpy
matplotlib
tqdm
pyyaml
tensorboard
h5py
pytest